refactor: UserDetails 구현 객체 분리
