+++ src/main/java/com/goormcoder/ieum/domain/MemberRole.java
@@ -2,10 +2,11 @@ package com.goormcoder.ieum.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
 
 @RequiredArgsConstructor
 @Getter
-public enum MemberRole {
+public enum MemberRole implements GrantedAuthority {
 
     ADMIN("admin"),
     USER("user"),
@@ -13,4 +14,8 @@ public enum MemberRole {
 
     private final String value;
 
+    @Override
+    public String getAuthority() {
+        return value;
+    }
 }
